src/backend/pyproject.toml
[tool.poetry]
name = "wb_adapter_view"
version = "0.0.1"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.11"
django = "^3.2.19"
environs = "10.3.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework = "^3.14.0"
structlog = "23.1.0"
psycopg2-binary = "^2.9.4"
celery = "^5.2.7"
redis = "^4.3.4"
requests = "^2.28.1"
tenacity = "^8.1.0"
dataclasses-json = "^0.5.7"
flower = "^1.2.0"
pydantic = "^1.10.8"
pyhumps = "^3.8.0"
django-debug-toolbar = "^4.1.0"
pytz = "^2023.3"
responses = "^0.23.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
autoflake = "^2.2.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
bpython = "^0.24"
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
create = true
in-project = true
